+++ server/.env
@@ -1,2 +1,3 @@
 STRIPE_KEY="sk_test_51NmfMlAZbxXHiVZzQCNJrB3DVqBF79hldMIqQWqKNXnnq9lkQBiDyBvFalMbxuSoVY7lnPkyPfEilRLR2lpE3ulE003RuorR9W"
 JWT_SECRET="ymsaeed1234"
+ENDPOINT_SECRET = "whsec_925f5e6e5a8acabe41114d465fcfa5b1e557567d324115618bd7842e6ad22f13"
