+++ server/routes/customers.route.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const { register, login, logout } = require("../controllers/customers.controller");
+const { register, login, logout, authorize } = require("../controllers/customers.controller");
 const jwt = require('jsonwebtoken');
 
 
@@ -8,28 +8,6 @@ const customerRouter = express
     .post("/customers/register", register)
     .post("/customers/login", login)
     .post("/customers/logout", logout)
-    .get('/customers/status', (req, res) => {
-        const token = req.cookies['auth-token'];
-        // console.log("Checking auth status. Token:", token);
-
-        if (token) {
-            try {
-                const decoded = jwt.verify(token, 'your_secret_key');
-                // console.log("Token verified successfully.");
-                res.status(200).send({ message: ` Hi, ${decoded.username}` });
-            } catch (err) {
-                console.log("Token verification failed:", err.message);
-                res.status(401).send();
-            }
-        } else {
-            // console.log("No token found in the request.");
-            res.status(401).send();
-        }
-    });
-
-
-
-
-// .get("/customers/authorize", authorize);
+    .get('/customers/status', authorize);
 
 module.exports = { customerRouter };
